refactor: 모임 만들기 훅 분리
